payslips

+++ src/main/java/com/mikepn/template/v1/dtos/request/deductions/UpdateDeductionDTO.java
@@ -2,8 +2,10 @@ package com.mikepn.template.v1.dtos.request.deductions;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class UpdateDeductionDTO {
     private String deductionName;
-    private double percentage;
+    private BigDecimal percentage;
 }
